Update createConditions method (#9)
* Update createConditions method

Update createConditions method to urlencode filter variables. I noticed some unanswered forum posts and needed to fix this quickly. Values passed for the 'filter' parameter were being appended as a raw JSON object and returning an empty error. I wrapped that in a urlEncode function to convert punctuation to hex and the API works. This keeps our code clean so we can keep passing arrays.

* Fix function name

urlEncode() switched to urlencode()